Fix the backlight default, the firework overflow, and the album-art fetch storm
--- the screen ---

TFT_BL_DEFAULT_INVERTED was true for the NM-TV-154, following the vendor
guide's TFT_BACKLIGHT_ON LOW. On real hardware that drives the panel backwards:
100% brightness writes duty 255, which inverts to 0 and switches the backlight
off, while a night level of 0 lights it fully. Confirmed on the device — "100%
turns it completely off".

It stayed hidden because a config saved before this default existed carried the
tick unset. A factory reset is what surfaces it, and the symptom then reads as
a bricked screen rather than a wrong setting: the device is dark, and the only
way back is a web UI you have no reason to believe is still running. Default is
now false for this board. The other three boards are untouched — I have no
measurement for them.

--- fireworks ---

The rewrite stored 8.8 fixed-point positions in int16_t. 240 << 8 does not fit:
x = 128 already overflows, so 44% of horizontal spawn positions and 25% of
vertical ones wrapped negative and the whole shell was culled on its first
frame. The pattern was mostly a black screen, which in a 180-second ambient
block reads as the device having switched off. Position is int32_t now.

I should have checked the arithmetic before shipping it; a five-line host
program found it immediately once I did.

--- album art ---

The device reported "HTTP -1, 3 covers offered", then fell into "player: bad
JSON" and stayed there. Three covers means the filter and the picker are
correct — that part was never broken. The failure is resource exhaustion:

  * The previous commit re-fetched the cover on every full repaint to stop it
    being wiped. A failed fetch left artFailed_ set, which forced a repaint,
    which retried the fetch. Each attempt burns a socket and a ~40 KB TLS
    arena, so every retry made the next one likelier to fail, and it took the
    poll's heap down with it. Working art to permanently broken in seconds.

  * renderAll now clears *around* the art box and leaves the cover in place
    when it is already correct, so a repaint costs no network at all. Fetches
    happen on a URL change and on a wake, which is when the cover is genuinely
    gone. wake() marks the box dirty, since another mode drawing is exactly
    what destroys it.

  * A failed fetch backs off 15 seconds rather than retrying on the next
    repaint, and a fetch will not start below 60 KB free heap — starting a
    handshake that cannot complete does not just fail, it strands the memory
    the poll needed.

  * The net lock wait dropped from 9 s to 2.5 s so a busy poll cannot stall the
    render loop.

The poll's JSON failures now name the actual DeserializationError instead of a
blanket "bad JSON"; a truncated stream and an exhausted heap want opposite
fixes.

--- layout ---

The artist line sat underneath the progress bar. 58 px under the cover has to
carry a title of up to two lines, the artist, two time labels and the bar, and
the spacing did not add up. Also the title picked size 2 for anything up to 26
characters against an 18-character line, so titles in between wrapped to a
second row and ran into the artist. Both fixed, and the layout arithmetic is
now checked by static_assert so the next adjustment cannot silently overlap.

+++ src/features/ambient/AmbientMode.h
@@ -73,7 +73,7 @@ class AmbientMode : public DisplayMode {
   // why the old bursts looked like a handful of darting specks: there was no
   // room between "barely moves" and "gone in three frames".
   struct Spark {
-    int16_t  x, y;         // 8.8 position
+    int32_t  x, y;         // 8.8 position — 32-bit: 240<<8 does not fit int16_t
     int16_t  vx, vy;       // 8.8 velocity, px per frame
     int16_t  px, py;       // last drawn pixel (whole pixels), -1 = not drawn
     uint8_t  life, age;    // frames remaining, frames lived
